+++ src/HistoryPointTable.h
@@ -1,6 +1,7 @@
 #ifndef __HISTORYPOINTTABLE_H__
 #define __HISTORYPOINTTABLE_H__
 
+#include "Types.h"
 #include "Point.h"
 #include <map>
 
@@ -11,17 +12,17 @@ class HistoryPointTable {
   public:
     HistoryPointTable(){};
 
-    void addNewPoint(long long memAddress, long long PC, long long numAccesses = 1, long long accessSize = 4,
+    void addNewPoint(int64_t memAddress, int64_t PC, int64_t numAccesses = 1, int64_t accessSize = 4,
                      MemAccessMode readOrWrite = READ);
 
-    bool doesPointExist(long long memAddress);
+    bool doesPointExist(int64_t memAddress);
 
-    list<Point *> getPoints(long long memAddress);
+    list<Point *> getPoints(int64_t memAddress);
 
     map<Interval, bool> getIntervalMap();
 
   private:
-    std::multimap<long long, Point *> myPoints;
+    std::multimap<int64_t, Point *> myPoints;
 };
 
 #endif
