+++ src/PendingStrideTable.h
@@ -1,6 +1,7 @@
 #ifndef PENDINGSTRIDETABLE_H
 #define PENDINGSTRIDETABLE_H
 
+#include "Types.h"
 #include "Stride.h"
 #include <iostream>
 #include <map>
@@ -14,18 +15,18 @@ class PendingStrideTable {
   public:
     PendingStrideTable(){};
 
-    void addNewStride(long long PC, Stride *stride);
+    void addNewStride(int64_t PC, Stride *stride);
 
-    void addNewStride(long long PCAddress, long long lowAddress, long long highAddress, long long strideLength,
-                      int numAccesses = 2, long long accessSize = 4, MemAccessMode readOrWrite = READ);
+    void addNewStride(int64_t PCAddress, int64_t lowAddress, int64_t highAddress, int64_t strideLength,
+                      int numAccesses = 2, int64_t accessSize = 4, MemAccessMode readOrWrite = READ);
 
-    bool doesStrideExist(long long PCAddress, long long memoryAddr, long long strideLength,
+    bool doesStrideExist(int64_t PCAddress, int64_t memoryAddr, int64_t strideLength,
                          MemAccessMode readOrWrite = READ);
 
-    bool updateExistingStride(long long PCAddress, long long memoryAddr, long long strideLength,
+    bool updateExistingStride(int64_t PCAddress, int64_t memoryAddr, int64_t strideLength,
                               MemAccessMode readOrWrite = READ);
 
-    bool killStride(long long PCAddress, long long memoryAddr, long long strideLength, MemAccessMode readOrWrite);
+    bool killStride(int64_t PCAddress, int64_t memoryAddr, int64_t strideLength, MemAccessMode readOrWrite);
 
     void print();
 
@@ -40,7 +41,7 @@ class PendingStrideTable {
         };
     };
 
-    std::multimap<long long, StrideEntry> myPendingStrides; // (PC, Stride*)
+    std::multimap<int64_t, StrideEntry> myPendingStrides; // (PC, Stride*)
 };
 
 #endif
